+++ db-backup/backup.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+pg_dump -h localhost -U postgres -W qa > qa-`date +%F`.sql
+